Separate `cmhn_compiler.py` module.
+ split `cmhn_compiler` into 3 modules:
    - `helptext` : command line driver / main module interface
    - `helptext_parser` : parser impl
    - `helptext_tests` : unit tests
    + `helptext_md` : markdown code generator
- The root history of `cmhn_compiler.py` will be kept in `helptext_parser.py` as that is where most of the work was done.

* Python tip for git repos is to split the module files as early as possible so that their history is maintained. Too late now, for the future.
* Wish to name the module in camelCase, but it seems py module naming rules insist on all lower alpha: https://peps.python.org/pep-0008/#package-and-module-names
